requirements.txt
opencv-python
numpy
tqdm
matplotlib
lib
psutil
numexpr
pillow
scikit-learn
tensorflow
tensorflow-metal
tensorflow-datasets
urllib3
keras-tuner
kaggle